@@ -3,6 +3,7 @@
   "version": "1.0.0",
   "description": "",
   "main": "server.js",
+  "type": "module",
   "scripts": {
     "serve": "node server.js NODE_ENV=production",
     "start": "NODE_ENV=development nodemon",
@@ -18,17 +19,17 @@
   "license": "MIT",
   "dependencies": {
     "cors": "^2.8.5",
-    "dotenv": "^8.2.0",
+    "dotenv": "^16.4.5",
     "ejs": "^3.1.6",
-    "eslint-config-prettier": "^8.3.0",
     "express": "^4.17.1",
-    "mongoose": "^5.11.15",
-    "node-fetch": "^2.6.1",
-    "prettier": "^2.3.1",
+    "mongoose": "^8.2.2",
+    "node-fetch": "^3.3.2",
+    "prettier": "^3.2.5",
     "shortid": "^2.2.16"
   },
   "devDependencies": {
-    "eslint": "^7.28.0",
-    "nodemon": "^2.0.7"
+    "eslint": "^8.57.0",
+    "eslint-config-prettier": "^9.1.0",
+    "nodemon": "^3.1.0"
   }
 }
