@@ -1,10 +1,10 @@
-const mongoose = require('mongoose');
+import { Schema, model } from 'mongoose';
 
-const originSchema = new mongoose.Schema({
+const originSchema = new Schema({
   endpoint: { type: String, required: true },
   shortid: { type: String, required: true },
   data: { type: Array, required: true },
 });
 
-const OriginModel = mongoose.model('Origin', originSchema);
-module.exports = OriginModel;
+const OriginModel = model('Origin', originSchema);
+export default OriginModel;
